requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.11
h11==0.14.0
idna==3.10
pydantic==2.10.6
pydantic-settings==2.1.0
pydantic_core==2.27.2
pytest==7.4.4
pytest-asyncio==0.23.5
httpx==0.26.0
pyyaml==6.0.1
psutil==5.9.5
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
